--- github_login/models.py
@@ -1,3 +1,0 @@
-from django.db import models
-
-# Create your models here.
