--- github_login/views.py
@@ -1,102 +1,0 @@
-# Create your views here.
-from json import JSONDecodeError
-
-from rest_framework import status
-import requests
-from allauth.socialaccount.models import SocialAccount
-from allauth.socialaccount.providers.github import views as github_view
-from allauth.socialaccount.providers.oauth2.client import OAuth2Client
-from django.conf import settings
-from django.http import JsonResponse
-from django.shortcuts import redirect
-from rest_auth.registration.views import SocialLoginView
-
-BASE_URL = 'http://localhost:8000/'
-GITHUB_CALLBACK_URI = BASE_URL + 'accounts/github/callback/'
-
-
-def github_login(request):
-    client_id = getattr(settings, 'SOCIAL_AUTH_GITHUB_KEY')
-    return redirect(
-        f"https://github.com/login/oauth/authorize?client_id={client_id}&redirect_uri={GITHUB_CALLBACK_URI}"
-    )
-
-
-def github_callback(request):
-    client_id = getattr(settings, 'SOCIAL_AUTH_GITHUB_CLIENT_ID')
-    client_secret = getattr(settings, 'SOCIAL_AUTH_GITHUB_SECRET')
-    code = request.GET.get('code')
-    """
-    Access Token Request
-    """
-    token_req = requests.post(
-        f"https://github.com/login/oauth/access_token?client_id={client_id}&client_secret={client_secret}&code={code}&accept=&json&redirect_uri={GITHUB_CALLBACK_URI}&response_type=code",
-        headers={'Accept': 'application/json'})
-    token_req_json = token_req.json()
-    error = token_req_json.get("error")
-    if error is not None:
-        raise JSONDecodeError(error)
-    access_token = token_req_json.get('access_token')
-    """
-    Email Request
-    """
-    user_req = requests.get(f"https://api.github.com/user",
-                            headers={"Authorization": f"Bearer {access_token}"})
-    user_json = user_req.json()
-    error = user_json.get("error")
-    if error is not None:
-        raise JSONDecodeError(error)
-    # print(user_json)
-    email = user_json.get("email")
-    """
-    Signup or Signin Request
-    """
-    try:
-        user = User.objects.get(email=email)
-        # 기존에 가입된 유저의 Provider가 github가 아니면 에러 발생, 맞으면 로그인
-        # 다른 SNS로 가입된 유저
-        social_user = SocialAccount.objects.get(user=user)
-        if social_user is None:
-            return JsonResponse({'err_msg': 'email exists but not social user'}, status=status.HTTP_400_BAD_REQUEST)
-        if social_user.provider != 'github':
-            return JsonResponse({'err_msg': 'no matching social type'}, status=status.HTTP_400_BAD_REQUEST)
-        # 기존에 github로 가입된 유저
-        data = {'access_token': access_token, 'code': code}
-        accept = requests.post(
-            f"{BASE_URL}accounts/github/login/finish/", data=data)
-        accept_status = accept.status_code
-        if accept_status != 200:
-            return JsonResponse({'err_msg': 'failed to signin'}, status=accept_status)
-        accept_json = accept.json()
-        accept_json.pop('user', None)
-        return JsonResponse(accept_json)
-    except User.DoesNotExist:
-        # 기존에 가입된 유저가 없으면 새로 가입
-        data = {'access_token': access_token, 'code': code}
-        accept = requests.post(
-            f"{BASE_URL}accounts/github/login/finish/", data=data)
-        accept_status = accept.status_code
-        if accept_status != 200:
-            return JsonResponse({'err_msg': 'failed to signup'}, status=accept_status)
-        # user의 pk, email, first name, last name과 Access Token, Refresh token 가져옴
-        accept_json = accept.json()
-        accept_json.pop('user', None)
-        return JsonResponse(accept_json)
-
-
-class GithubLogin(SocialLoginView):
-    """
-    If it's not working
-    You need to customize GitHubOAuth2Adapter
-    use header instead of params
-    -------------------
-    def complete_login(self, request, app, token, **kwargs):
-        params = {'access_token': token.token}
-TO
-def complete_login(self, request, app, token, **kwargs):
-        headers = {'Authorization': 'Bearer {0}'.format(token.token)}
-    -------------------
-    """
-    adapter_class = github_view.GitHubOAuth2Adapter
-    callback_url = GITHUB_CALLBACK_URI
-    client_class = OAuth2Client
