+++ user_custom/admin.py
@@ -1,3 +1,17 @@
+# Register your models here.
 from django.contrib import admin
 
-# Register your models here.
+from .models import User
+from django.contrib import admin
+from rest_framework_simplejwt.token_blacklist.admin import OutstandingTokenAdmin
+from rest_framework_simplejwt.token_blacklist.models import OutstandingToken
+
+
+class CustomOutstandingTokenAdmin(OutstandingTokenAdmin):
+    def has_delete_permission(self, *args, **kwargs):
+        return True
+
+
+admin.site.unregister(OutstandingToken)
+# admin.site.register(OutstandingToken, CustomOutstandingTokenAdmin)
+admin.site.register(User)
