+++ user_custom/serializers.py
@@ -1,10 +1,12 @@
 from rest_framework import serializers
+from rest_framework_simplejwt.tokens import RefreshToken
 
-from user_custom.models import CustomUser
 
+class CustomTokenRefreshSerializer(serializers.Serializer):
+    refresh_token = serializers.CharField()
 
-class UserSerializer(serializers.ModelSerializer):
+    def validate(self, attrs):
+        refresh = RefreshToken(attrs['refresh_token'])
+        data = {'access_token': str(refresh.access_token)}
 
-    class Meta:
-        model = CustomUser
-        fields = ['id', 'email', 'first_name', 'last_name', 'spouse_name', 'date_of_birth']
+        return data
